public_mirrors/vetka-taskboard/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sse-starlette>=1.6.0
pydantic>=2.0
